Fixed warning

+++ LambdaEngine/Source/Resources/ResourceLoader.cpp
@@ -878,7 +878,7 @@ namespace LambdaEngine
 		s_pComputeCommandList->BindDescriptorSetCompute(s_pCubeMapGenDescriptorSet, s_pCubeMapGenPipelineLayout, 0);
 
 		constexpr uint32 THREADGROUP_SIZE = 8;
-		const uint32 dispatchSize = AlignUp(size, THREADGROUP_SIZE) / THREADGROUP_SIZE;
+		const uint32 dispatchSize = uint32(AlignUp(size, THREADGROUP_SIZE) / THREADGROUP_SIZE);
 		s_pComputeCommandList->Dispatch(size, size, 6);
 
 		s_pComputeCommandList->TransitionBarrier(
